refactor(data-model): split `BaseFabricInstance` out of `FabricInstance`
`FabricInstance` is now purely abstract: it pins the protocol surface
(`[DECONSTRUCT]`, `[DEEP_FREEZE]`, `[IS_DEEP_FROZEN]`, `deepClone`,
`shallowClone`) but provides no implementations. Shared template-method
scaffolding lives on a new `BaseFabricInstance` subclass in
`fabric-instances/`, alongside `ExplicitTagValue`. Today that covers
`shallowClone()` / `shallowUnfrozenClone()`; the existing `deepClone()`
TODO now points at `BaseFabricInstance` as the future home of an
analogous base implementation.

`ExplicitTagValue` and `FabricNativeWrapper` reparent to
`BaseFabricInstance`; concrete leaf classes inherit through them with no
changes. `packages/api/index.ts` mirror is left alone -- the api
`FabricInstance` keeps `shallowClone` as its external protocol surface,
and `BaseFabricInstance` stays an internal implementation detail.
